backend/skillswap/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
django==4.2.11
djangorestframework==3.15.1
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1

celery==5.3.6
redis==5.0.1
django-celery-results==2.5.1
django-celery-beat==2.6.0


psycopg2-binary==2.9.9


python-dotenv==1.0.1
requests==2.31.0


openai==2.30.0

langchain-core==0.2.38
langchain-groq==0.1.9

chromadb>=0.4.0








pdfplumber==0.10.3
channels

channels-redis

django-allauth
dj-rest-auth
google-auth>=2.0.0


torch
sentence-transformers==2.6.1